(undone) rating

+++ src/main/java/project/fashionecommerce/backend/fashionecommerceproject/controller/admin/manager/review/models/ReviewRequest.java
@@ -9,7 +9,7 @@ public record ReviewRequest(
         String orderId,
         String content,
         @NotNull
-        Double rating,
+        Long rating,
         List<String> images,
         @NotNull
         Boolean isDeleted,
